src/requirements.txt
apparmor==3.0.3
appdirs==1.4.3
application-utility==1.3.2
apsw==3.36.0.post1
asgiref==3.2.10
attrs==21.2.0
autopep8==1.5.4
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
btrfsutil==5.13.1
CacheControl==0.12.6
cached-property==1.5.2
cchardet==2.1.7
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2020.6.20
cffi==1.14.6
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.8
css-parser==1.0.6
cssselect==1.1.0
cupshelpers==1.0
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-richtextfield==1.6
django-storages==1.10.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
dnspython==1.16.0
docker==5.0.2
docker-compose==1.29.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
dropbox==10.7.0
evdev==1.4.0
feedparser==5.2.1
filelock==3.0.12
gunicorn==20.0.4
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
idna==2.10
ifaddr==0.1.7
importlib-metadata==4.8.1
ipaddr==2.2.0
jeepney==0.7.1
jsonfield==3.1.0
jsonschema==3.2.0
keyutils==0.6
LibAppArmor==3.0.3
libtorrent===1.2.14-build-libtorrent-rasterbar-src-libtorrent-rasterbar-1.2.14-bindings-python
lockfile==0.12.2
louis==3.19.0
lxml==4.6.1
Markdown==3.3.4
markdown2==2.3.10
mechanize==0.4.5
more-itertools==8.8.0
msgpack==0.6.2
mypy==0.910
mypy-extensions==0.4.3
netifaces==0.11.0
netsnmp-python==1.0a1
nftables==0.1
npyscreen==4.10.5
ordered-set==4.0.2
packaging==20.3
pacman-mirrors==4.21.5
paramiko==2.7.2
pep517==0.8.2
Pillow==8.0.1
platformdirs==2.2.0
ply==3.11
progress==1.5
psutil==5.8.0
psycopg2==2.9.1
py7zr==0.11.3
pycairo==1.20.1
pychm==0.8.6
pycodestyle==2.6.0
pycparser==2.20
pycryptodome==3.10.1
pycups==2.0.1
pycurl==7.43.0.6
pyenchant==3.2.1
Pygments==2.10.0
PyGObject==3.40.1
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.6
PyQt5==5.15.4
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.4
pyquery==1.4.1
pyrsistent==0.18.0
pysmbc==1.0.23
python-dateutil==2.8.2
python-dotenv==0.19.0
pytoml==0.1.21
pytz==2020.1
pyxdg==0.27
PyYAML==5.4.1
rados==2.0.0
rbd==2.0.0
readtime==1.1.1
regex==2021.7.6
reportlab==3.6.1
requests==2.24.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
swapper==1.1.2.post1
team==1.0
texttable==1.6.4
toml==0.10.2
tomli==1.2.1
types-pytz==2021.1.2
types-PyYAML==5.4.10
typing-extensions==3.10.0.2
udiskie==2.3.3
ufw==0.36
unrardll==0.1.4
urllib3==1.25.11
virtualenv==20.7.2
webencodings==0.5.1
websocket-client==0.59.0
whitenoise==5.2.0
youtube-dl==2021.6.6
zeroconf==0.29.0
zipp==3.5.0
